requirements.txt
streamlit
openai
faiss-cpu
numpy
transformers
torch
Pillow
aiohttp
beautifulsoup4
